requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Cache==0.12
Flask-SQLAlchemy==1.0
Flask-Script==0.6.2
Flask-WTF==0.9.2
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
SQLAlchemy==0.8.2
Sphinx==1.2b1
WTForms==1.0.5
Werkzeug==0.9.4
blinker==1.3
coverage==3.6
cssmin==0.1.4
docutils==0.11
flake8==2.0
itsdangerous==0.23
jsmin
mccabe==0.2.1
pep8==1.4.6
py==1.4.15
pyflakes==0.7.3
pytest==2.3.5
webassets==0.8
wsgiref==0.1.2
flask-pymongo
flask-restful
flask-login
